final project feedback
[This comment is a supplement to the feedback in email.]

Generally a pull request is meant to be accepted into a code base. This one should not be. I’m including code changes that are improvements, but also fragments and comments that are meant to be read as commentary, but not added to the repo.

The code comments also include changes to spacing, because this is what my editor does. This is somewhat useful because it demonstrates standard style, but not the main lesson – so look at a couple of these, but then look for lines with more substantive changes.

General comments:
- Some class and method docs, especially in algo
- Some very clear code, too. Sophisticated use of @staticmethod
- remove unused files from the repo. This makes it easier to maintain a project. (For example, global search searches only current code; it prevents the maintainer getting confused and editing a stale file.) Create a git branch or git tag to mark the last revisions that contain code that is then deleted